move social manager to auth package
