+++ app/src/main/java/org/stepik/android/adaptive/api/auth/AuthRepository.kt
@@ -2,7 +2,6 @@ package org.stepik.android.adaptive.api.auth
 
 import io.reactivex.Single
 import org.stepik.android.adaptive.api.RegistrationResponse
-import org.stepik.android.adaptive.api.login.SocialManager
 import org.stepik.android.adaptive.data.model.RegistrationUser
 import retrofit2.Response
